refactor: move nlp features to natural_language_processing.py
