+++ api/terraform/python/openai_api/lambda_openai_v2/lambda_handler.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=no-member
-# FIX NOTE: pylint is not recognizing the 'openai' module.
+# pylint: disable=R0801
 """
 written by: Lawrence McDaniel
             https://lawrencemcdaniel.com/
